tree_collapse.js: Make collapsing more predictable
Previously a subtree would collapse on clicking anywhere in its
parent `<li>` element. This made it difficult for a user to
predict whether it would collapse or not. I changed it so that it
collapses only when the `.index-heading` element is clicked.
This is predictable because `.index-heading`'s style changes when
mouse is hovered over it.